requirements.txt
ruamel.yaml>=0.17.21
typer==0.7.0
pydantic >=2, <3
pyyaml==6.0
openpyxl >= 3.1.2, == 3.*
schemapack==2.0.0a3